+++ src/reports/InstancesIdReport.js
@@ -1,26 +1,28 @@
-import { get } from 'lodash';
 import moment from 'moment';
+import { noop } from 'lodash';
 
 import { exportCsv } from '@folio/stripes/util';
 
-class InstancesIdReport {
-  parse(records) {
-    return records.map(record => {
-      const toCSV = {
-        id: get(record, 'id'),
-      };
+import { isTestEnv } from '../utils';
 
-      return toCSV;
-    });
+class InstancesIdReport {
+  constructor(fileNamePrefix) {
+    this.fileNamePrefix = fileNamePrefix;
   }
 
-  toCSV(records = []) {
-    const parsedRecords = this.parse(records);
+  parse(records, recordFinder = record => record) {
+    return records.map(record => ({ id: recordFinder(record) }));
+  }
 
-    exportCsv(parsedRecords, {
+  toCSV(records, recordFinder) {
+    const parsedRecords = this.parse(records, recordFinder);
+    const fileTitle = {
       header: false,
-      filename: 'SearchInstanceUUIDs' + moment().format(),
-    });
+      filename: `${this.fileNamePrefix}${moment().format()}`,
+    };
+    const generateReport = !isTestEnv() ? exportCsv : noop;
+
+    generateReport(parsedRecords, fileTitle);
   }
 }