+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 * Upgraded to create-inventory plugin v2.0.0.
 * Include missing item-status values. Refs UIIN-1385.
 * Add show selected records action menu. Refs UIIN-1375.
+* Implement quick export of instances in MARC Bib output format. Refs UIIN-1382.
 
 ## [5.0.1](https://github.com/folio-org/ui-inventory/tree/v5.0.1) (2020-10-15)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v5.0.0...v5.0.1)